requirements.txt
pandas
numpy
scikit-learn
sentence-transformers
catboost
fastapi
uvicorn
pydantic
joblib
pydantic
torch
tf-keras
catboost
optuna
shap
matplotlib
seaborn
matplotlib
seaborn
wordcloud
pytest
streamlit
